+++ src/main/java/net/mtrop/doom/object/GraphicObject.java
@@ -7,6 +7,9 @@
  ******************************************************************************/
 package net.mtrop.doom.object;
 
+import net.mtrop.doom.graphics.Flat;
+import net.mtrop.doom.graphics.Picture;
+
 /**
  * Interface for graphic data.
  * @author Matthew Tropiano
@@ -34,4 +37,29 @@ public interface GraphicObject
 	 */
 	public int getHeight();
 
+	/**
+	 * Gets the pixel data at a location in the graphic.
+	 * <p>If this graphic is an indexed color graphic (i.e. {@link Flat} or {@link Picture}), this
+	 * will return a palette index value from 0 to 255, or {@link Picture#PIXEL_TRANSLUCENT} if this graphic has translucent pixels.
+	 * <p>For full-color graphics, this returns an ARGB integer value representing the pixel color in RGB space (with Alpha).  
+	 * @param x	graphic x-coordinate.
+	 * @param y	graphic y-coordinate.
+	 * @return a palette index value from 0 to 255, {@link Picture#PIXEL_TRANSLUCENT} if translucent, or an ARGB value.
+	 * @throws ArrayIndexOutOfBoundsException if the provided coordinates is outside the graphic.
+	 */
+	public int getPixel(int x, int y);
+
+	/**
+	 * Sets the pixel data at a location in the graphic.
+	 * <p>For indexed color graphics, valid values are in the range of -1 to 255, 
+	 * with 0 to 254 being palette indexes and {@link Picture#PIXEL_TRANSLUCENT} / 255 being translucent pixels (if supported).
+	 * <p>For full-color graphics, the value is an ARGB integer value representing the pixel color in RGB space (with Alpha).
+	 * @param x	picture x-coordinate.
+	 * @param y	picture y-coordinate.
+	 * @param value	the value to set.
+	 * @throws IllegalArgumentException if the value is outside a valid range.
+	 * @throws ArrayIndexOutOfBoundsException if the provided coordinates is outside the graphic.
+	 */
+	public void setPixel(int x, int y, int value);
+	
 }